Refactor proxy normalization and add affine transforms

+++ __init__.py
@@ -19,9 +19,16 @@ class ProxyNormalization(nn.Module):
         beta = create_channelwise_variable(self.y, 0.0)
         gamma = create_channelwise_variable(self.y, 1.0)
 
+        #Affine transform =  gamma * n + beta
+
+        #affine transform on y and apply activation function
         z = self.activation_fn(gamma * self.y.add_(beta))
+
+        #Create a proxy distribution of y
         proxy_y = torch.tensor(uniformly_sampled_gaussian(self.n_samples), dtype=self.y.dtype)
-        proxy_y = torch.randn_like(self.y)
+        proxy_y = torch.randn_like(self.y, dtype=self.y.dtype)
+        
+        #Affine transform on proxy of y and apply activation function
         proxy_z = self.activation_fn(gamma * proxy_y.add_(beta))
 
         proxy_mean = torch.mean(proxy_z, dim=0, keepdim=True)
